Add Home Assistant MQTT detail
Fan control in native Tasmota when using TuyaMCU is not well implemented yet so it required customised MQTT configurations as per another Arlec fan already in the templates (https://templates.blakadder.com/arlec_ATF4601HA.html).

The light control added here also ensures the light works correctly with the way the physical remote controls the brightness.  That is; it keeps them in sync so that there is no behaviour confusion when using HA vs the remote.